Temporarily revert r315602. Bring back the definition for the GCC __nonnull() attribute.
Old versions of GCC, including the version installed the latest 11-stable
snapshot with pkg(8), still carry the old attributes.

The issue is easily fixed by rebuilding GCC but there is no need to cause
havoc in our user base. The definition by itself is harmless but it should
be removed again in the near future.


git-svn-id: https://svn.freebsd.org/base/stable/11@316779 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f