MFC 313772:
Load the new sp_el0 with interrupts disabled in fork_trampoline. If an
interrupt arrives in fork_trampoline after sp_el0 was written we may then
switch to a new thread, enter userland so change this stack pointer, then
return to this code with the wrong value. This fixes this case by moving
the load of sp_el0 until after interrupts have been disabled.


git-svn-id: https://svn.freebsd.org/base/stable/11@317147 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f