Attempt to determine the modes in which 8-bit wide characters are actually 9 wide.
I only need this to improve the mouse cursor, but it has always been
needed to select and/or adjust fonts.

This is complicated because there are no standard parameter tables
giving this bit of information directly, and the device register bit
giving the information can't be trusted even if it is read from the
hardware.  Use a heuristic to guess if the device register can be
trusted.  (The device register is normally read from the BIOS mode
table, but on my system where the device register is wrong, the mode
table doesn't match the hardware and is not used; the device registers
are used in this case.)


git-svn-id: https://svn.freebsd.org/base/head@317190 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f